Renamed Iteration method

+++ branches/nodefs/grails-app/domain/org/agiletracking/Iteration.groovy
@@ -70,7 +70,7 @@ class Iteration extends ItemContainer {
 		return days
 	}
 	
-	Integer calculatePointsPerDay()
+	Double calculatePointsPerDay()
 	{
 		def ret = 0
